chore(xtask): count every cfg(test) item, not only the first
`production_lines` stopped at the first column-zero `#[cfg(test)]` and
treated everything after it as tests. A `#[cfg(test)] use` near the top
of a file therefore hid the whole file from the 1,200-line cap:
`runstate.rs` measured 46 lines for a real 1,015, `control_socket/mod.rs`
35 for 759, `gemini.rs` 5 for 521, and eleven more.

The count now skips each column-zero test item (its attribute stack,
then a `;` one-liner or a bracketed item closing at column zero) and
resumes after it. Indented attributes still count as production, which
over-reports a handful of files in the safe direction. Under the honest
count nothing crosses the cap, so the fix lands without a split - except
`commands/update.rs`, which sat at 1,196 and now has `update/detect.rs`
(install-method detection) and `update/migrate.rs` (config migrations)
beside it, a pure move with the public names re-exported.

The four stray column-zero test items that triggered the old failure
(`gemini.rs`, `openai.rs`, mcp `transport/http.rs`, mcp `auth/mod.rs`)
move into their files' test modules, and `lint_gaps` now checks
`xtask/Cargo.toml` too.
