+++ crates/leviath-providers/src/openai.rs
@@ -4,8 +4,6 @@ use crate::openai_compat::{
     OpenAiSseStream, TokenLimitField, build_openai_request_body_with, parse_openai_response,
     send_chat_request, temperature_refused, tools_refused_over_reasoning_effort,
 };
-#[cfg(test)]
-use crate::provider::FinishReason;
 use crate::provider::{
     InferenceRequest, InferenceResponse, LimitsSource, ModelCapabilities, ModelCapabilityOverride,
     ModelInfo, Provider, ProviderConfig, ProviderError, Result, StreamChunk,
@@ -450,6 +448,7 @@ impl Provider for OpenAIProvider {
 
 #[cfg(test)]
 mod tests {
+    use crate::provider::FinishReason;
 
     /// Config beats the shipped table, and an unconfigured model still gets the
     /// published rate rather than falling to unpriced.
